a bit of refactoring. added the get to see the user's events route. wrote on each route-file the routes you can access there. deleted some unused routes.

+++ routes/user.routes.js
@@ -6,7 +6,12 @@ const fileUploader = require("../config/cloudinary.config"); //DOWNLOAD?
 const User = require("../models/User.model");
 const Comment = require("../models/Comment.model");
 
-/* http://localhost:5005/api/users */
+/* WHAT CAN YOU FIND HERE?
+- GET ROUTE TO SEE ALL USERS IN APP
+- GET ROUTE TO SEE DETAILS OF CERTAIN USER 
+- PUT ROUTE TO EDIT PROFILE
+- DELETE ROUTE TO DELETE USER
+*/
 
 router.get("/users", (req, res, next) => {
   User.find()
@@ -20,7 +25,7 @@ router.get("/users/:id", (req, res, next) => {
   const { id } = req.params;
 
   User.findById(id)
-    .populate("following followers  myEvents")
+    .populate("following followers myEvents")
     .populate({
       path: "receivedComments",
       populate: { path: "commentBy" },
@@ -28,11 +33,13 @@ router.get("/users/:id", (req, res, next) => {
     .then((user) => {
       res.status(200).json({ user });
     })
-    .catch((err) => res.status(500).json({ errorMessage: error.message }));
+    .catch((err) => res.status(500).json({ errorMessage: "error.message" }));
 });
 
-router.put("/users/:id", (req, res, next) => {
+router.put("/users/:id", fileUploader.single("image"), (req, res, next) => {
   const { id } = req.params;
+  const { _id } = req.payload;
+
   const {
     image,
     username,
@@ -43,6 +50,13 @@ router.put("/users/:id", (req, res, next) => {
     skills,
     links,
   } = req.body;
+
+  if (id !== _id) {
+    res.status(400).json({
+      errorMessage: "You can't edit someone's profile who's not you.",
+    });
+    return;
+  }
 
   User.findByIdAndUpdate(
     id,
@@ -67,10 +81,19 @@ router.put("/users/:id", (req, res, next) => {
 
 router.delete("/users/:id", async (req, res, next) => {
   const { id } = req.params;
+  const { _id } = req.payload;
+
+  if (id !== _id) {
+    res.status(400).json({
+      errorMessage: "You can't delete someone's profile who's not you.",
+    });
+    return;
+  }
 
   let deletedUser = await User.findByIdAndRemove(id);
   await Event.deleteMany({ _id: { $in: deletedUser.myEvents } });
   await Comment.deleteMany({ _id: { $in: deletedUser.myComments } });
+  await Comment.deleteMany({ _id: { $in: deletedUser.receivedComments } });
 
   res.status(200).json({ user });
